rolled back code for mail function
