+++ application/templates/login.html
@@ -19,7 +19,7 @@
       </div>
     </div>
     <div class="login-box well">
-      <form action="studDash.html" method="POST">
+      <form action="studDash" method="POST">
         <div class="form-group">
           <label id="up">Username</label>
           <input type="text" name="username" class="form-control" value="" , placeholder="eg. johnsmith">
